refactor(engine): move common aria-related code into shared package
