+++ packages/@lwc/shared/src/aria.ts
@@ -5,9 +5,18 @@
  * For full license text, see the LICENSE file in the repo root or https://opensource.org/licenses/MIT
  */
 
-// Global Aria and Role Properties derived from ARIA and Role Attributes.
-// https://wicg.github.io/aom/spec/aria-reflection.html
-export const AriaPropertyNames = [
+import { create, forEach, StringReplace, StringToLowerCase } from './language';
+
+/**
+ * According to the following list, there are 48 aria attributes of which two (ariaDropEffect and
+ * ariaGrabbed) are deprecated:
+ * https://www.w3.org/TR/wai-aria-1.1/#x6-6-definitions-of-states-and-properties-all-aria-attributes
+ *
+ * The above list of 46 aria attributes is consistent with the following resources:
+ * https://github.com/w3c/aria/pull/708/files#diff-eacf331f0ffc35d4b482f1d15a887d3bR11060
+ * https://wicg.github.io/aom/spec/aria-reflection.html
+ */
+const AriaPropertyNames = [
     'ariaActiveDescendant',
     'ariaAtomic',
     'ariaAutoComplete',
@@ -56,3 +65,77 @@ export const AriaPropertyNames = [
     'ariaValueText',
     'role',
 ];
+
+// We should eventually use the typescript Pick utility to reuse ambient Aria property types when
+// the typescript interface for Element is updated to include them.
+export interface AccessibleElementProperties {
+    ariaActiveDescendant: string | null;
+    ariaAtomic: string | null;
+    ariaAutoComplete: string | null;
+    ariaBusy: string | null;
+    ariaChecked: string | null;
+    ariaColCount: string | null;
+    ariaColIndex: string | null;
+    ariaColSpan: string | null;
+    ariaControls: string | null;
+    ariaCurrent: string | null;
+    ariaDescribedBy: string | null;
+    ariaDetails: string | null;
+    ariaDisabled: string | null;
+    ariaErrorMessage: string | null;
+    ariaExpanded: string | null;
+    ariaFlowTo: string | null;
+    ariaHasPopup: string | null;
+    ariaHidden: string | null;
+    ariaInvalid: string | null;
+    ariaKeyShortcuts: string | null;
+    ariaLabel: string | null;
+    ariaLabelledBy: string | null;
+    ariaLevel: string | null;
+    ariaLive: string | null;
+    ariaModal: string | null;
+    ariaMultiLine: string | null;
+    ariaMultiSelectable: string | null;
+    ariaOrientation: string | null;
+    ariaOwns: string | null;
+    ariaPlaceholder: string | null;
+    ariaPosInSet: string | null;
+    ariaPressed: string | null;
+    ariaReadOnly: string | null;
+    ariaRelevant: string | null;
+    ariaRequired: string | null;
+    ariaRoleDescription: string | null;
+    ariaRowCount: string | null;
+    ariaRowIndex: string | null;
+    ariaRowSpan: string | null;
+    ariaSelected: string | null;
+    ariaSetSize: string | null;
+    ariaSort: string | null;
+    ariaValueMax: string | null;
+    ariaValueMin: string | null;
+    ariaValueNow: string | null;
+    ariaValueText: string | null;
+    role: string | null;
+}
+
+const AttrNameToPropNameMap: Record<string, string> = create(null);
+const PropNameToAttrNameMap: Record<string, string> = create(null);
+
+// Synthetic creation of all AOM property descriptors for Custom Elements
+forEach.call(AriaPropertyNames, (propName) => {
+    // Typescript infers the wrong function type for this particular overloaded method:
+    // https://github.com/Microsoft/TypeScript/issues/27972
+    // @ts-ignore type-mismatch
+    const attrName = StringToLowerCase.call(StringReplace.call(propName, /^aria/, 'aria-'));
+    AttrNameToPropNameMap[attrName] = propName;
+    PropNameToAttrNameMap[propName] = attrName;
+});
+
+export {
+    AttrNameToPropNameMap as AriaAttrNameToPropNameMap,
+    PropNameToAttrNameMap as AriaPropNameToAttrNameMap,
+};
+
+export function isAriaAttribute(attrName: string) {
+    return Boolean(AttrNameToPropNameMap[attrName]);
+}
