+++ packages/@lwc/template-compiler/src/parser/constants.ts
@@ -4,6 +4,8 @@
  * SPDX-License-Identifier: MIT
  * For full license text, see the LICENSE file in the repo root or https://opensource.org/licenses/MIT
  */
+import { AriaAttrNameToPropNameMap } from '@lwc/shared';
+
 import { HTML_ATTRIBUTE_ELEMENT_MAP } from './utils/html-element-attributes';
 import { HTML_ELEMENTS, HTML_VOID_ELEMENTS } from './utils/html-elements';
 
@@ -43,7 +45,7 @@ const ATTRIBUTE_NAME_CHAR = [
     '\\u200D\\u2070-\\u218F\\u2C00-\\u2FEF\\u3001-\\uD7FF\\uF900-\\uFDCF\\uFDF0-\\uFFFD',
     '\\-.0-9\\u00B7\\u0300-\\u036F\\u203F-\\u2040',
 ].join('');
-export const ARIA_RE = new RegExp('^(aria)-[' + ATTRIBUTE_NAME_CHAR + ']*$');
+
 export const DATA_RE = new RegExp('^(data)-[' + ATTRIBUTE_NAME_CHAR + ']*$');
 
 export const SUPPORTED_SVG_TAGS = new Set([
@@ -161,8 +163,6 @@ export const DASHED_TAGNAME_ELEMENT_SET = new Set([
 ]);
 
 export const ATTRS_PROPS_TRANFORMS: { [name: string]: string } = {
-    // These have to be quoted because aria- attributes below are quoted
-    // Our linting doesn't like to mix quoted and un-quoted property names
     accesskey: 'accessKey',
     readonly: 'readOnly',
     tabindex: 'tabIndex',
@@ -179,58 +179,7 @@ export const ATTRS_PROPS_TRANFORMS: { [name: string]: string } = {
     novalidate: 'noValidate',
     usemap: 'useMap',
     for: 'htmlFor',
-    // According to the following list, there are 48 aria attributes (of which 2 are deprecated):
-    // https://www.w3.org/TR/wai-aria-1.1/#x6-6-definitions-of-states-and-properties-all-aria-attributes
-    // This list of 46 non-deprecated aria attributes is consistent with the following proposal:
-    // https://github.com/w3c/aria/pull/708/files#diff-eacf331f0ffc35d4b482f1d15a887d3bR11060
-    'aria-activedescendant': 'ariaActiveDescendant',
-    'aria-atomic': 'ariaAtomic',
-    'aria-autocomplete': 'ariaAutoComplete',
-    'aria-busy': 'ariaBusy',
-    'aria-checked': 'ariaChecked',
-    'aria-colcount': 'ariaColCount',
-    'aria-colindex': 'ariaColIndex',
-    'aria-colspan': 'ariaColSpan',
-    'aria-controls': 'ariaControls',
-    'aria-current': 'ariaCurrent',
-    'aria-describedby': 'ariaDescribedBy',
-    'aria-details': 'ariaDetails',
-    'aria-disabled': 'ariaDisabled',
-    //  'aria-dropeffect': 'ariaDropEffect', /* Deprecated in ARIA 1.1 */
-    'aria-errormessage': 'ariaErrorMessage',
-    'aria-expanded': 'ariaExpanded',
-    'aria-flowto': 'ariaFlowTo',
-    //  'aria-grabbed': 'ariaGrabbed', /* Deprecated in ARIA 1.1 */
-    'aria-haspopup': 'ariaHasPopup',
-    'aria-hidden': 'ariaHidden',
-    'aria-invalid': 'ariaInvalid',
-    'aria-keyshortcuts': 'ariaKeyShortcuts',
-    'aria-label': 'ariaLabel',
-    'aria-labelledby': 'ariaLabelledBy',
-    'aria-level': 'ariaLevel',
-    'aria-live': 'ariaLive',
-    'aria-modal': 'ariaModal',
-    'aria-multiline': 'ariaMultiLine',
-    'aria-multiselectable': 'ariaMultiSelectable',
-    'aria-orientation': 'ariaOrientation',
-    'aria-owns': 'ariaOwns',
-    'aria-placeholder': 'ariaPlaceholder',
-    'aria-posinset': 'ariaPosInSet',
-    'aria-pressed': 'ariaPressed',
-    'aria-readonly': 'ariaReadOnly',
-    'aria-relevant': 'ariaRelevant',
-    'aria-required': 'ariaRequired',
-    'aria-roledescription': 'ariaRoleDescription',
-    'aria-rowcount': 'ariaRowCount',
-    'aria-rowindex': 'ariaRowIndex',
-    'aria-rowspan': 'ariaRowSpan',
-    'aria-selected': 'ariaSelected',
-    'aria-setsize': 'ariaSetSize',
-    'aria-sort': 'ariaSort',
-    'aria-valuemax': 'ariaValueMax',
-    'aria-valuemin': 'ariaValueMin',
-    'aria-valuenow': 'ariaValueNow',
-    'aria-valuetext': 'ariaValueText',
+    ...AriaAttrNameToPropNameMap,
 };
 
 export const DISALLOWED_HTML_TAGS = new Set(['base', 'link', 'meta', 'script', 'title']);
